pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gdsync"
version = "0.1.0"
description = "Export and sync Google Workspace files from Drive into multiple formats"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: System :: Archiving :: Backup",
]
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.1.0",
    "markdownify>=0.11.0",
    "pyyaml>=6.0",
]

[project.scripts]
gdsync = "gdsync.cli:main"

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]

[tool.setuptools.packages.find]
include = ["gdsync*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
